Refactor testing of Device
